+++ e2e/test-env/src/client/RicosTestApp/editorPlugins.js
@@ -10,6 +10,7 @@ import { pluginHashtag } from 'wix-rich-content-plugin-hashtag';
 import { pluginHeadings } from 'wix-rich-content-plugin-headings';
 import { pluginSpoiler } from 'wix-rich-content-plugin-spoiler';
 import { pluginAccordion } from 'wix-rich-content-plugin-accordion';
+import { pluginTable } from 'wix-rich-content-plugin-table';
 import { pluginHeadersMarkdown } from 'wix-rich-content-plugin-headers-markdown';
 import { pluginHtml } from 'wix-rich-content-plugin-html';
 import { pluginImage } from 'wix-rich-content-plugin-image';
@@ -117,6 +118,16 @@ const plugins = {
       pluginLink().createPlugin,
       pluginCodeBlock().createPlugin,
       pluginImage().createPlugin,
+    ],
+  }),
+  table: pluginTable({
+    innerRCEPlugins: [
+      pluginTextColor(configs.textColor).createPlugin,
+      pluginTextHighlight(configs.textHighlight).createPlugin,
+      pluginIndent().createPlugin,
+      pluginLineSpacing().createPlugin,
+      pluginLink().createPlugin,
+      pluginCodeBlock().createPlugin,
     ],
   }),
   verticalEmbed: pluginVerticalEmbed(configs.verticalEmbed),