+++ packages/editor-common/web/src/Utils/draftUtils.ts
@@ -11,7 +11,7 @@ import {
   EditorChangeType,
 } from '@wix/draft-js';
 
-import { flatMap, findIndex, findLastIndex, countBy, debounce, times } from 'lodash';
+import { cloneDeepWith, flatMap, findIndex, findLastIndex, countBy, debounce, times } from 'lodash';
 import { TEXT_TYPES } from '../consts';
 import { RelValue, AnchorTarget } from 'wix-rich-content-common';
 import { Optional } from 'utility-types';
@@ -25,6 +25,9 @@ type LinkDataUrl = {
 };
 
 type LinkData = LinkDataUrl & { anchor?: string };
+
+const isEditorState = value => value?.getCurrentContent && value;
+export const cloneDeepWithoutEditorState = obj => cloneDeepWith(obj, isEditorState);
 
 export function createSelection({
   blockKey,
@@ -335,7 +338,11 @@ export const createBlockAndFocus = (editorState: EditorState, data, pluginType: 
 export const createBlock = (editorState: EditorState, data, type: string) => {
   const currentEditorState = editorState;
   const contentState = currentEditorState.getCurrentContent();
-  const contentStateWithEntity = contentState.createEntity(type, 'IMMUTABLE', { ...data });
+  const contentStateWithEntity = contentState.createEntity(
+    type,
+    'IMMUTABLE',
+    cloneDeepWithoutEditorState(data)
+  );
   const entityKey = contentStateWithEntity.getLastCreatedEntityKey();
   const newEditorState = AtomicBlockUtils.insertAtomicBlock(currentEditorState, entityKey, ' ');
   const recentlyCreatedKey = newEditorState.getSelection().getAnchorKey();
@@ -696,4 +703,63 @@ export function deleteCharacterBeforeCursor(editorState: EditorState) {
     newState = EditorState.push(editorState, newContentState, 'delete-string' as EditorChangeType);
   }
   return newState;
+}
+
+export function isPluginFocused(block: ContentBlock, selection: SelectionState) {
+  return block.getKey() === selection.getAnchorKey();
+}
+
+export function isCursorAtFirstLine(editorState: EditorState) {
+  const selection = editorState.getSelection();
+  return (
+    selection.isCollapsed() &&
+    editorState
+      .getCurrentContent()
+      .getBlockMap()
+      .first()
+      .getKey() === selection.getFocusKey()
+  );
+}
+
+export function isCursorAtStartOfContent(editorState: EditorState) {
+  const isStartOfLine = editorState.getSelection().getFocusOffset() === 0;
+  return isStartOfLine && isCursorAtFirstLine(editorState);
+}
+
+export function isCursorAtLastLine(editorState: EditorState) {
+  const selection = editorState.getSelection();
+  return (
+    selection.isCollapsed() &&
+    editorState
+      .getCurrentContent()
+      .getBlockMap()
+      .last()
+      .getKey() === selection.getFocusKey()
+  );
+}
+
+export function isCursorAtEndOfContent(editorState: EditorState) {
+  const selection = editorState.getSelection();
+  const lastBlock = editorState
+    .getCurrentContent()
+    .getBlockMap()
+    .last();
+  const isEndOfLine = selection.getFocusOffset() >= lastBlock.getText().length;
+  return isEndOfLine && isCursorAtLastLine(editorState);
+}
+
+export function selectAllContent(editorState, forceSelection) {
+  const currentContent = editorState.getCurrentContent();
+  const selection = editorState.getSelection().merge({
+    anchorKey: currentContent.getFirstBlock().getKey(),
+    anchorOffset: 0,
+
+    focusOffset: currentContent.getLastBlock().getText().length,
+    focusKey: currentContent.getLastBlock().getKey(),
+  });
+  const setSelectionFunction = forceSelection
+    ? EditorState.forceSelection
+    : EditorState.acceptSelection;
+  const newEditorState = setSelectionFunction(editorState, selection);
+  return newEditorState;
 }