+++ packages/editor/web/src/RichContentEditor/Toolbars/SideToolbar/createSideToolbar.jsx
@@ -3,6 +3,7 @@ import SideToolbar from './SideToolbar';
 import AddPluginFloatingToolbar from './AddPluginFloatingToolbar';
 import { decorateComponentWithProps } from 'wix-rich-content-editor-common';
 import { simplePubsub } from 'wix-rich-content-common';
+import { getPluginMenuTheme } from './utils';
 
 const createSideToolbar = (data = {}) => {
   const {
@@ -58,16 +59,7 @@ export default ({
   config,
   addPluginMenuConfig,
 }) => {
-  const { buttonStyles, ...rest } = theme;
-  const toolbarButtonTheme = {
-    buttonStyles: {
-      button: buttonStyles.sideToolbarButton,
-      buttonWrapper: buttonStyles.sideToolbarButton_wrapper,
-      icon: buttonStyles.sideToolbarButton_icon,
-      label: buttonStyles.sideToolbarButton_label,
-    },
-    ...rest,
-  };
+  const toolbarButtonTheme = getPluginMenuTheme(theme, isMobile);
   return createSideToolbar({
     offset,
     theme,