+++ packages/plugin-table/web/src/viewer.ts
@@ -1,12 +1,6 @@
-/*
-  This module exports the required CreatePluginFunction for RicosViewer.
-  If your plugin uses decorations, then make sure to uncomment 'decorator'.
-  (Please find examples of usage in other plugins)
-*/
-
 import { typeMapper } from './typeMapper';
+import { theme, DEFAULTS } from './defaults';
 import { TABLE_TYPE, TablePluginViewerConfig } from './types';
-import { DEFAULTS, theme } from './defaults';
 import { ViewerPluginCreator } from 'wix-rich-content-common';
 export { TABLE_TYPE, typeMapper as tableTypeMapper };